requirements.txt
numpy==1.21.4
pandas==1.3.4
streamlit==1.2.0
streamlit_drawable_canvas==0.8.0
tensorflow-cpu==2.7.0
matplotlib-inline==0.1.3
matplotlib==3.4.3